implementing bids changes
