+++ arcana/dataset/__init__.py
@@ -1,4 +1,5 @@
-from .particular import Dataset, Field, DatasetCollection, FieldCollection
+from .particular import Dataset, Field
+from .collection import DatasetCollection, FieldCollection
 from .spec import DatasetSpec, FieldSpec, BaseSpec
 from .base import BaseField, BaseDataset, BaseDatum
 from .match import DatasetMatch, FieldMatch, BaseMatch
